update license key handling.
