+++ electron.js
@@ -16,7 +16,7 @@ const Config = require('electron-config');
 var appCfg = {
   mainWindow: { width: 800, height: 665, x:null, y:null },
   helpWindow: { width: 700, height: 500 },
-  systemWindow: { width: 390, height: 400 },
+  systemWindow: { width: 390, height: 530 },
   audioServVer: 'webaudioapi', // html5audio or webaudioapi
   showMsgPane: true,
   acceptFirstMouse: false,
