+++ main.html
@@ -12,6 +12,8 @@
   <script src="node_modules/angular-input-highlight/angular-input-highlight.js"></script>
   <script src="node_modules/intro.js/minified/intro.min.js"></script>
   <script src="js/models.js"></script>
+  <script src="js/services.intro.js"></script>
+  <script src="js/services.license.js"></script>
   <script src="js/services.js"></script>
   <script src="js/app.js"></script>
   <script type="text/javascript">
