+++ system.html
@@ -7,6 +7,7 @@
   <link rel="stylesheet" href="css/system.css">
   <script src="node_modules/photon/dist/template-app/js/menu.js" charset="utf-8"></script>
   <script src="node_modules/angular/angular.min.js"></script>
+  <script src="js/services.license.js"></script>
   <script src="js/system.js"></script>
   <script type="text/javascript">
     document.ondragover = document.ondrop = function(e) {
@@ -61,6 +62,7 @@
       <div class="form-group">
         <label>AquesTalk10 使用ラインセンスキー</label>
         <input type="text" class="form-control" id="aq10UseKey" ng-model="aq10UseKey" placeholder="XXXX-XXXX-XXXX-XXXX">
+        <small>AquesTalk10の音声を再生するには、AquesTalk10の使用ライセンスキー(Mac用)を設定してください</small>
       </div>
     </div>
   </form>
